refactor: rename BlockAction to PoSAction
